backend/requirements.txt
Django==3.2
djangorestframework==3.12.4
mysqlclient==2.0.3
django-cors-headers==3.7.0
djangorestframework-simplejwt==4.7.2
